fix(common): scope b20 variant selectors
