+++ crates/common/precompiles/src/b20_security/dispatch.rs
@@ -109,12 +109,6 @@ impl<S: SecurityAccounting, P: Policy> B20SecurityToken<S, P> {
             C::symbol(_) => self.accounting.symbol()?.abi_encode().into(),
             C::decimals(_) => U256::from(self.accounting.decimals()?).abi_encode().into(),
             C::totalSupply(_) => self.accounting.total_supply()?.abi_encode().into(),
-            C::minimumRedeemable(_) => self.accounting.minimum_redeemable()?.abi_encode().into(),
-            C::currency(_) => self.accounting.currency()?.abi_encode().into(),
-            // securityIdentifier also caught by IB20Security above; repeated for exhaustiveness.
-            C::securityIdentifier(c) => {
-                self.accounting.security_identifier(&c.identifierType)?.abi_encode().into()
-            }
             C::balanceOf(c) => self.accounting.balance_of(c.account)?.abi_encode().into(),
             C::allowance(c) => self.accounting.allowance(c.owner, c.spender)?.abi_encode().into(),
             C::supplyCap(_) => self.accounting.supply_cap()?.abi_encode().into(),
@@ -376,6 +370,7 @@ impl<S: SecurityAccounting, P: Policy> B20SecurityToken<S, P> {
             }
 
             // --- Minimum redeemable (security version, in shares) ---
+            SC::minimumRedeemable(_) => self.accounting.minimum_redeemable()?.abi_encode().into(),
             SC::updateMinimumRedeemable(c) => {
                 self.accounting_mut().set_minimum_redeemable(c.newMinimumRedeemable)?;
                 self.accounting_mut().emit_event(
