+++ crates/common/precompiles/src/b20_stablecoin/accounting.rs
@@ -11,6 +11,9 @@ use crate::TokenAccounting;
 /// Only [`super::B20StablecoinToken`] requires this bound; default and security
 /// tokens use the base [`TokenAccounting`] port exclusively.
 pub trait StablecoinAccounting: TokenAccounting {
+    /// Returns the stablecoin currency identifier.
+    fn currency(&self) -> Result<String>;
+
     /// Writes the currency identifier. Called once by the factory at creation.
     fn set_currency(&mut self, currency: String) -> Result<()>;
 }
