+++ crates/common/precompiles/src/b20_stablecoin/dispatch.rs
@@ -3,14 +3,18 @@
 //! Dispatches the full `IB20` selector set using B-20 stablecoin activation.
 //! All logic mirrors `B20Token::inner_with_privilege` exactly; the only
 //! distinction is the activation guard and the `StablecoinAccounting` bound
-//! that provides `currency()` from EVM slot 11.
+//! that provides `currency()` from the stablecoin extension namespace.
 
 use alloy_primitives::{Bytes, U256};
-use alloy_sol_types::SolValue;
+use alloy_sol_types::{SolInterface, SolValue};
 use base_precompile_storage::{BasePrecompileError, IntoPrecompileResult, StorageCtx};
 use revm::precompile::PrecompileResult;
 
-use super::{B20StablecoinToken, accounting::StablecoinAccounting};
+use super::{
+    B20StablecoinToken,
+    abi::{IB20Stablecoin, IB20Stablecoin::IB20StablecoinCalls as SC},
+    accounting::StablecoinAccounting,
+};
 use crate::{
     ActivationFeature, ActivationRegistryStorage, B20TokenRole, Burnable, Configurable,
     IB20::{self, IB20Calls as C},
@@ -53,6 +57,10 @@ impl<S: StablecoinAccounting, P: Policy> B20StablecoinToken<S, P> {
         ActivationRegistryStorage::new(ctx)
             .ensure_activated(ActivationFeature::B20Stablecoin.id())?;
 
+        if let Ok(call) = IB20Stablecoin::IB20StablecoinCalls::abi_decode(calldata) {
+            return self.handle_stablecoin_call(call);
+        }
+
         let call = decode_precompile_call!(calldata, IB20::IB20Calls);
 
         let encoded: Bytes = match call {
@@ -61,11 +69,6 @@ impl<S: StablecoinAccounting, P: Policy> B20StablecoinToken<S, P> {
             C::symbol(_) => self.accounting.symbol()?.abi_encode().into(),
             C::decimals(_) => U256::from(self.accounting.decimals()?).abi_encode().into(),
             C::totalSupply(_) => self.accounting.total_supply()?.abi_encode().into(),
-            C::minimumRedeemable(_) => self.accounting.minimum_redeemable()?.abi_encode().into(),
-            C::currency(_) => self.accounting.currency()?.abi_encode().into(),
-            C::securityIdentifier(c) => {
-                self.accounting.security_identifier(&c.identifierType)?.abi_encode().into()
-            }
             C::balanceOf(c) => self.accounting.balance_of(c.account)?.abi_encode().into(),
             C::allowance(c) => self.accounting.allowance(c.owner, c.spender)?.abi_encode().into(),
             C::supplyCap(_) => self.accounting.supply_cap()?.abi_encode().into(),
@@ -231,6 +234,13 @@ impl<S: StablecoinAccounting, P: Policy> B20StablecoinToken<S, P> {
                 )?;
                 Bytes::new()
             }
+        };
+        Ok(encoded)
+    }
+
+    fn handle_stablecoin_call(&self, call: SC) -> base_precompile_storage::Result<Bytes> {
+        let encoded: Bytes = match call {
+            SC::currency(_) => self.accounting.currency()?.abi_encode().into(),
         };
         Ok(encoded)
     }
