+++ crates/common/precompiles/src/b20_stablecoin/token.rs
@@ -13,7 +13,7 @@ use crate::{
 /// EVM precompile for the stablecoin B-20 variant.
 ///
 /// Mirrors the structure of [`crate::B20Token`] but requires `S: StablecoinAccounting`
-/// so the dispatch layer can read `currency()` from storage. All inherited
+/// so the dispatch layer can read `currency()` from stablecoin storage. All inherited
 /// `IB20` capability traits are wired in identically.
 #[derive(Debug, Clone)]
 pub struct B20StablecoinToken<S: StablecoinAccounting, P: Policy> {
