+++ crates/common/precompiles/src/lib.rs
@@ -33,7 +33,7 @@ pub use common::{InMemoryPolicy, InMemoryTokenAccounting, TestStablecoinToken, T
 mod b20;
 pub use b20::{
     B20CoreStorage, B20PausableFeature, B20PolicyType, B20Token, B20TokenPrecompile,
-    B20TokenStorage, IB20, POLICY_ALWAYS_ALLOW, POLICY_ALWAYS_BLOCK,
+    B20TokenStorage, IB20,
 };
 
 mod b20_security;
